Added pytest for functional tests, reading executor config from env
