+++ .env.example
@@ -1,0 +1,5 @@
+executor_username=
+executor_conda_env=
+# terraform outputs
+executor_hostname=
+executor_ssh_key_file=
