+++ .github/workflows/tests.yml
@@ -89,7 +89,7 @@ jobs:
 
       - name: Run tests
         run: |
-          PYTHONPATH=$PWD/tests pytest -vv tests/ --cov=covalent_ssh_plugin
+          PYTHONPATH=$PWD/tests pytest -vv tests/ --ignore=./tests/functional_tests  --cov=covalent_ssh_plugin
 
       - name: Generate coverage report
         run: coverage xml -o coverage.xml
