+++ tests/functional_tests/requirements.txt
@@ -1,2 +1,3 @@
 numpy==1.23.2
+python-dotenv==0.21.0
 scikit-learn==1.1.2
