--- tests/functional_tests/terraform_output.py
@@ -1,33 +1,0 @@
-import json
-import os
-import subprocess
-import sys
-import traceback
-
-TERRAFORM_OUTPUTS = {}
-
-try:
-    terraform_dir = os.getenv("TF_DIR")
-    proc = subprocess.run(
-        [
-            "terraform",
-            f"-chdir={terraform_dir}",
-            "output",
-            "-json",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    TERRAFORM_OUTPUTS = json.loads(proc.stdout.decode())
-except Exception:
-    traceback.print_exc()
-
-print("Terraform output:")
-print(TERRAFORM_OUTPUTS)
-
-
-def get(key: str, default):
-    try:
-        return TERRAFORM_OUTPUTS[key]["value"]
-    except KeyError:
-        return default
